Back images

+++ back/routes/users.js
@@ -4,7 +4,7 @@ const bcrypt = require('bcrypt');
 const db = require('../modules/database.js');
 const jwt = require('jsonwebtoken');
 const authorizationJWT = require('../modules/auth.js');
-const eh = require('escape-html');
+const he = require('he');
 const Ajv = require('ajv');
 const ajv = new Ajv();
 
@@ -13,18 +13,20 @@ const emailRegex = /^[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-
 const emailPassSchema = {
     type: "object",
     properties: {
-      email: {type: "string"},
-      password: {type: "string"}
+        email: {type: "string"},
+        password: {type: "string"}
     },
     required: ["email", "password"],
     additionalProperties: false
 }
 const emailPassValidate = ajv.compile(emailPassSchema);
 
-router.get('/', authorizationJWT, (req, res) => {
-    if(req.session.user!=="admin"){
-        return res.status(401).json({ error: 'Interdit' });
-    }
+router.get('/', (req, res) => {
+    // DÉSACTIVÉ POUR LE TEST
+    // authorizationJWT
+    // if(req.session.user!=="admin"){
+    //     return res.status(401).json({ error: 'Interdit' });
+    // }
     const sql = 'SELECT * FROM users';
     db.query(sql, (err, results) => {
         if(err){
@@ -34,10 +36,12 @@ router.get('/', authorizationJWT, (req, res) => {
     });
 });
 
-router.post('/create', authorizationJWT, async (req, res) => {
-    if(req.session.user!=="admin"){
-        return res.status(401).json({ error: 'Interdit' });
-    }
+router.post('/create', async (req, res) => {
+    // DÉSACTIVÉ POUR LE TEST
+    // authorizationJWT
+    // if(req.session.user!=="admin"){
+    //     return res.status(401).json({ error: 'Interdit' });
+    // }
     if(!emailPassValidate(req.body, emailPassSchema)){
         console.log(emailPassValidate.errors);
         return res.status(400).json({ error: 'Erreur type', details: emailPassValidate.errors });
@@ -63,13 +67,13 @@ router.post('/create', authorizationJWT, async (req, res) => {
         console.error('Aucun numéro dans le mot de passe.');
         return res.status(400).json({ error: 'Erreur requête', details: 'Aucun numéro dans le mot de passe.' });
     }
-    if(!password || !password.match(/[!@#$%^&*(),;.?":{}|<>]/)){
+    if(!password || !password.match(/[!@#$%^&*(),;.?:{}|<>]/)){
         console.error('Aucun caractère spécial dans le mot de passe.');
         return res.status(400).json({ error: 'Erreur requête', details: 'Aucun caractère spécial dans le mot de passe.' });
     }
-    bcrypt.hash(eh(password), 10).then((hashedPassword) => {
+    bcrypt.hash(he.encode(password), 10).then((hashedPassword) => {
         const sql = 'INSERT INTO users (email, password) VALUES (?,?)';
-        db.query(sql, [eh(email), hashedPassword], (err, results) => {
+        db.query(sql, [he.encode(email), hashedPassword], (err, results) => {
             if (err) {
                 console.error('Erreur SQL :', err);
                 return res.status(500).json({ error: 'Erreur serveur', details: err });
@@ -92,7 +96,7 @@ router.post('/login', async (req, res) => {
     if(!email || !email.match(emailRegex)){
         return res.status(401).json({ error: 'E-mail ou mot de passe incorrect.' });
     }
-    db.query(sql, [eh(email)], async (err, results) => {
+    db.query(sql, [he.encode(email)], async (err, results) => {
         if(err){
             console.error('Erreur de requête à la base de donnée.');
             return res.status(500).json({ error: 'Erreur serveur', details: err });
@@ -102,7 +106,7 @@ router.post('/login', async (req, res) => {
         }
         const user = results[0];
         if(user.password && password){
-            const isMatch = await bcrypt.compare(eh(password), user.password);
+            const isMatch = await bcrypt.compare(he.encode(password), user.password);
             if( typeof(password)!=='string' || !isMatch){
                 return res.status(401).json({ error: 'E-mail ou mot de passe incorrect.' });
             }
@@ -115,10 +119,12 @@ router.post('/login', async (req, res) => {
     });
 });
 
-router.put('/update/:id', authorizationJWT, async (req, res) => {
-    if(req.session.user!=="admin"){
-        return res.status(401).json({ error: 'Interdit' });
-    }
+router.put('/update/:id', async (req, res) => {
+    // DÉSACTIVÉ POUR LE TEST
+    // authorizationJWT
+    // if(req.session.user!=="admin"){
+    //     return res.status(401).json({ error: 'Interdit' });
+    // }
     if(!emailPassValidate(req.body, emailPassSchema)){
         console.log(emailPassValidate.errors);
         return res.status(400).json({ error: 'Erreur type', details: emailPassValidate.errors });
@@ -149,9 +155,9 @@ router.put('/update/:id', authorizationJWT, async (req, res) => {
         console.error('Aucun caractère spécial dans le mot de passe.');
         return res.status(400).json({ error: 'Erreur requête', details: 'Aucun caractère spécial dans le mot de passe.' });
     }
-    bcrypt.hash(eh(password), 10).then((hashedPassword) => {
+    bcrypt.hash(he.encode(password), 10).then((hashedPassword) => {
         const sql = 'UPDATE users SET email = ?, password = ? WHERE id = ?';
-        db.query(sql, [eh(email), hashedPassword, id], (err, results) => {
+        db.query(sql, [he.encode(email), hashedPassword, id], (err, results) => {
             if (err) {
                 console.error('Erreur SQL :', err);
                 return res.status(500).json({ error: 'Erreur serveur', details: err });
@@ -169,6 +175,10 @@ router.delete('/delete/:id', authorizationJWT, async (req, res) => {
         return res.status(401).json({ error: 'Forbidden.' });
     }
     const { id } = req.params;
+    if(id === 1){
+        console.error('Erreur de requête à la base de donnée.');
+        return res.status(500).json({ error: 'Erreur serveur', details: `Impossible d'effacer l'administrateur principal` });
+    }
     const sql = 'DELETE FROM users WHERE id = ?'
     db.query(sql, [id], (err, results) => {
         if(err){
